Fixed the main CoreData issue!! Now just working on correctly representing the data we store
