--- HyprRun/CoreData/CoreDataStack.swift
@@ -1,41 +1,0 @@
-//
-//  CoreDataStack.swift
-//  HyprRun
-//
-//  Created by Katie Lin on 11/11/22.
-//  Reference: https://www.kodeco.com/553-how-to-make-an-app-like-runkeeper-part-1
-//
-
-import CoreData
-
-class CoreDataStack {
-  
-  static let persistentContainer: NSPersistentContainer = {
-    let container = NSPersistentContainer(name: "HyprRun")
-    container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-      if let error = error as NSError? {
-        fatalError("Unresolved error \(error), \(error.userInfo)")
-      }
-    })
-    return container
-  }()
-  
-  static var context: NSManagedObjectContext {
-    return persistentContainer.viewContext
-  }
-  
-  class func saveContext() {
-    let context = persistentContainer.viewContext
-    
-    guard context.hasChanges else {
-      return
-    }
-    
-    do {
-      try context.save()
-    } catch {
-      let nserror = error as NSError
-      fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-    }
-  }
-}
