+++ HyprRun/HyprRunApp.swift
@@ -23,7 +23,6 @@ struct HyprRunApp: App {
     var body: some Scene {
         WindowGroup {
 					ContentView(isAuthorized: $isAuthorized)
-            .environment(\.managedObjectContext, CoreDataStack.context)
 						.environmentObject(spotify)
         }
     }
